files from templet
these are all the files from the templet that i purcheds minus two java script files because it says wow thats alot of files. my hope is to upload these now then the last two after thes are loaded. plan b would be to weed out some of the files that i will not be needing to make room.